+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3spatial (development version)
+
 # mlr3spatial 0.6.1
 
 * compatibility: mlr3 1.2.0 removed the data format argument