+++ NEWS.md
@@ -1,6 +1,6 @@
 # mlr3spatial (development version)
 
-* feat: `predict_spatial()` supports probability predictions now. Train a classification learner with `predict_type = "prob"` to get one raster layer or one vector column per class (#88, #89).
+* feat: `predict_spatial()` supports probability predictions now. Train a classification learner with `predict_type = "prob"` to get one raster layer or one vector column per class.
 
 # mlr3spatial 0.6.1